Calcul coupe minimale corrigé et proposition 1 de calcul de robustesse

+++ src/robotmodel.h
@@ -34,13 +34,13 @@ void StepTarget(double * TargetPosition,
 
 /* Step positions and velocities
  */
-void Step(phase_t * OutputPhase, phase_t * GPSPhase, phase_t * GPSDelayedPhase,
-        phase_t * PhaseData, unit_model_params_t * UnitParams, int WhichTarget,
-        flocking_model_params_t * FlockingParams, sit_parameters_t * SitParams,
-        vizmode_params_t * VizParams, int TimeStepLooped, int TimeStepReal,
-        bool CountCollisions, bool * ConditionsReset, int *Collisions,
-        bool * AgentsInDanger, double *WindVelocityVector, double *Accelerations,
-        double ** TargetsArray, double **Polygons, node **Hull, int Verbose, int *pRobustness);
+void Step(phase_t *OutputPhase, phase_t *GPSPhase, phase_t *GPSDelayedPhase,
+          phase_t *PhaseData, unit_model_params_t *UnitParams, int WhichTarget,
+          flocking_model_params_t *FlockingParams, sit_parameters_t *SitParams,
+          vizmode_params_t *VizParams, int TimeStepLooped, int TimeStepReal,
+          bool CountCollisions, bool *ConditionsReset, int *Collisions,
+          bool *AgentsInDanger, double *WindVelocityVector, double *Accelerations,
+          double **TargetsArray, double **Polygons, node **Hull, int Verbose, float *pRobustness);
 
 /*  Initialization and killing
  */
